separate component logic out and add message component
